refactor(PiecePrintFormat): 클래스명/변수명 변경
